backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
requests>=2.28.0
stripe>=6.0.0
PyJWT>=2.8.0
